+++ packages/angular/src/injectLazyRead/index.ts
@@ -123,16 +123,6 @@ export function createInjectLazyRead<
       }
     });
 
-    const reset = () => {
-      if (currentQueryKey) {
-        stateManager.deleteCache(currentQueryKey);
-      }
-
-      dataSignal.set(undefined);
-      errorSignal.set(undefined);
-      loadingSignal.set(false);
-    };
-
     const abort = () => {
       currentController?.abort();
     };
@@ -246,7 +236,6 @@ export function createInjectLazyRead<
       data: dataSignal as Signal<TData | undefined>,
       error: errorSignal as Signal<TError | undefined>,
       loading: loadingSignal,
-      reset,
       abort,
     };